Move invertVccLutPinsToGndPins() into RouterHelperSupport
